+++ model/purchase.gen.go
@@ -8,7 +8,7 @@ const TableNamePurchase = "Purchase"
 
 // Purchase mapped from table <Purchase>
 type Purchase struct {
-	ID          int32   `gorm:"column:id;primaryKey" json:"id"`
+	ID          int64   `gorm:"column:id;primaryKey" json:"id"`
 	CreatedAt   string  `gorm:"column:created_at;not null" json:"created_at"`
 	UpdatedAt   string  `gorm:"column:updated_at;not null" json:"updated_at"`
 	AmountMoney float64 `gorm:"column:amount_money;not null" json:"amount_money"`
@@ -16,14 +16,14 @@ type Purchase struct {
 	Status      string  `gorm:"column:status;not null" json:"status"`
 	SourceType  string  `gorm:"column:source_type;not null" json:"source_type"`
 	PurchasedItems []PurchasedItem `gorm:"foreignKey:PurchaseID;references:PurchaseID"`
-	CustomerID  int32   `gorm:"column:customerId" json:"customerId"`
+	CustomerID  int64   `gorm:"column:customerId" json:"customerId"`
 	PurchaseID  string  `gorm:"column:purchase_id;not null" json:"purchase_id"`
 	ProductType string  `gorm:"column:product_type" json:"product_type"`
-	LocationID  int32   `gorm:"column:locationId;not null" json:"locationId"`
-	ItemsID     int32   `gorm:"column:itemsId" json:"itemsId"`
+	LocationID  int64   `gorm:"column:locationId;not null" json:"locationId"`
+	ItemsID     int64   `gorm:"column:itemsId" json:"itemsId"`
 	Location   StoreLocation `gorm:"foreignKey:LocationID;references:ID"`
 	AttributedCampaign Campaign `gorm:"foreignKey:AttributedCampaignID;references:ID"`
-	AttributedCampaignID int32 `gorm:"column:attributedCampaignId;not null" json:"attributedCampaignId"`
+	AttributedCampaignID int64 `gorm:"column:attributedCampaignId;not null" json:"attributedCampaignId"`
 	Customer  Customer `gorm:"foreignKey:CustomerID;references:ID"`
 }
