+++ model/store_location.gen.go
@@ -8,7 +8,7 @@ const TableNameStoreLocation = "store_location"
 
 // StoreLocation mapped from table <store_location>
 type StoreLocation struct {
-	ID                           int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID                           int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	Name                         string `gorm:"column:name;not null" json:"name"`
 	AddressLine1                 string `gorm:"column:address_line_1;not null" json:"address_line_1"`
 	Locality                     string `gorm:"column:locality;not null" json:"locality"`
@@ -24,7 +24,7 @@ type StoreLocation struct {
 	Currency                     string `gorm:"column:currency;not null" json:"currency"`
 	PhoneNumber                  string `gorm:"column:phone_number" json:"phone_number"`
 	BusinessName                 string `gorm:"column:business_name;not null" json:"business_name"`
-	OrganizationID               int32  `gorm:"column:organizationId;not null" json:"organizationId"`
+	OrganizationID               int64  `gorm:"column:organizationId;not null" json:"organizationId"`
 	PosID                        string `gorm:"column:pos_id;not null" json:"pos_id"`
 	Type                         string `gorm:"column:type" json:"type"`
 }
